1. 헥사고날 아키텍쳐에 맞춘 패키지 구조 리팩토링 2. 서비스 인터페이스 생성 및 구현체로 리팩토링 3. 작성자 서비스 0.0.3 업데이트
