+++ docker/docker-compose.yml
@@ -102,7 +102,7 @@ services:
       - backend
 
   writerservice:
-    image: msa/writer-service:0.0.2-SNAPSHOT
+    image: msa/writer-service:0.0.3-SNAPSHOT
     ports:
       - "8082:8082"
     environment:
