+++ writer-service/src/main/java/com/dasd412/api/writerservice/application/service/FindWriterService.java
@@ -1,0 +1,7 @@
+package com.dasd412.api.writerservice.application.service;
+
+import com.dasd412.api.writerservice.domain.writer.Writer;
+
+public interface FindWriterService extends WriterService{
+    Writer findWriterById(Long id);
+}
