+++ writer-service/src/main/java/com/dasd412/api/writerservice/application/service/SaveWriterService.java
@@ -1,0 +1,6 @@
+package com.dasd412.api.writerservice.application.service;
+
+public interface SaveWriterService extends WriterService {
+
+    Long saveWriterInTest(String name, String email);
+}
