+++ writer-service/src/main/java/com/dasd412/api/writerservice/application/service/UpdateWriterService.java
@@ -1,0 +1,8 @@
+package com.dasd412.api.writerservice.application.service;
+
+public interface UpdateWriterService {
+
+    void attachRelationWithDiary(Long writerId, Long diaryId);
+
+    void detachRelationWithDiary(Long writerId, Long diaryId);
+}
