+++ writer-service/src/main/java/com/dasd412/api/writerservice/application/service/WriterService.java
@@ -1,0 +1,4 @@
+package com.dasd412.api.writerservice.application.service;
+
+public interface WriterService {
+}
