+++ writer-service/src/main/java/com/dasd412/api/writerservice/application/service/impl/DeleteWriterServiceImpl.java
@@ -1,0 +1,20 @@
+package com.dasd412.api.writerservice.application.service.impl;
+
+import com.dasd412.api.writerservice.adapter.out.persistence.WriterRepository;
+import com.dasd412.api.writerservice.application.service.DeleteWriterService;
+import com.dasd412.api.writerservice.application.service.WriterService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
+
+@Service
+public class DeleteWriterServiceImpl implements DeleteWriterService {
+
+    private final WriterRepository writerRepository;
+
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+
+    public DeleteWriterServiceImpl(WriterRepository writerRepository) {
+        this.writerRepository = writerRepository;
+    }
+}
