+++ writer-service/src/main/java/com/dasd412/api/writerservice/common/config/QueryDSLConfiguration.java
@@ -1,0 +1,24 @@
+package com.dasd412.api.writerservice.common.config;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+@Configuration
+public class QueryDSLConfiguration {
+
+    /**
+     * Querydsl 작업할 때는 JPAQueryFactory Bean이 필요하다.
+     * 그리고 생성할 때는 EntityManager를 주입해야 한다.
+     */
+    @PersistenceContext
+    private EntityManager entityManager;
+
+    @Bean
+    public JPAQueryFactory jpaQueryFactory() {
+        return new JPAQueryFactory(entityManager);
+    }
+}
