+++ writer-service/src/main/java/com/dasd412/api/writerservice/common/utils/UserContextHolder.java
@@ -1,0 +1,28 @@
+package com.dasd412.api.writerservice.common.utils;
+
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+public class UserContextHolder {
+
+    private static final ThreadLocal<UserContext> userContext = new ThreadLocal<>();
+
+    public static UserContext getContext() {
+        UserContext context = userContext.get();
+
+        if (context == null) {
+            context = createEmptyContext();
+            setUserContext(context);
+        }
+        return userContext.get();
+    }
+
+    public static void setUserContext(UserContext context) {
+        checkNotNull(context,"User context must be not null!!");
+        userContext.set(context);
+    }
+
+    private static UserContext createEmptyContext() {
+        return new UserContext();
+    }
+}
