--- writer-service/src/main/java/com/dasd412/api/writerservice/config/RestTemplateConfiguration.java
@@ -1,26 +1,0 @@
-package com.dasd412.api.writerservice.config;
-
-import com.dasd412.api.writerservice.utils.UserContextInterceptor;
-import org.springframework.cloud.client.loadbalancer.LoadBalanced;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.ClientHttpRequestInterceptor;
-import org.springframework.web.client.RestTemplate;
-
-import java.util.List;
-
-@Configuration
-public class RestTemplateConfiguration {
-
-    @LoadBalanced
-    @Bean
-    public RestTemplate getRestTemplate(){
-        RestTemplate template=new RestTemplate();
-
-        List<ClientHttpRequestInterceptor> interceptors=template.getInterceptors();
-
-        interceptors.add(new UserContextInterceptor());
-        template.setInterceptors(interceptors);
-        return template;
-    }
-}
