--- writer-service/src/main/java/com/dasd412/api/writerservice/controller/ApiError.java
@@ -1,36 +1,0 @@
-package com.dasd412.api.writerservice.controller;
-
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-import org.springframework.http.HttpStatus;
-
-public class ApiError {
-    private final String message;
-
-    private final int status;
-
-    ApiError(Throwable throwable, HttpStatus status) {
-        this(throwable.getMessage(), status);
-    }
-
-    ApiError(String message, HttpStatus status) {
-        this.message = message;
-        this.status = status.value();
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE)
-                .append("message", message)
-                .append("status", status)
-                .toString();
-    }
-}
