--- writer-service/src/main/java/com/dasd412/api/writerservice/domain/writer/WriterRepository.java
@@ -1,8 +1,0 @@
-package com.dasd412.api.writerservice.domain.writer;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WriterRepository extends JpaRepository<Writer, Long>, WriterRepositoryCustom {
-}
