--- writer-service/src/main/java/com/dasd412/api/writerservice/domain/writer/WriterRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.dasd412.api.writerservice.domain.writer;
-
-public interface WriterRepositoryCustom {
-}
