--- writer-service/src/main/java/com/dasd412/api/writerservice/exception/NotSupportedKafkaMessageException.java
@@ -1,15 +1,0 @@
-package com.dasd412.api.writerservice.exception;
-
-public class NotSupportedKafkaMessageException extends IllegalArgumentException{
-    public NotSupportedKafkaMessageException(String s) {
-        super(s);
-    }
-
-    public NotSupportedKafkaMessageException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public NotSupportedKafkaMessageException(Throwable cause) {
-        super(cause);
-    }
-}
