--- writer-service/src/main/java/com/dasd412/api/writerservice/message/DiaryActionEnum.java
@@ -1,15 +1,0 @@
-package com.dasd412.api.writerservice.message;
-
-public enum DiaryActionEnum {
-    CREATED,
-    DELETED,
-    NOT_SUPPORTED;
-
-    public static Enum compare(String string) {
-        try {
-            return valueOf(string.toUpperCase());
-        } catch (IllegalArgumentException e) {
-            return NOT_SUPPORTED;
-        }
-    }
-}
