--- writer-service/src/main/java/com/dasd412/api/writerservice/utils/UserContextInterceptor.java
@@ -1,22 +1,0 @@
-package com.dasd412.api.writerservice.utils;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpRequest;
-import org.springframework.http.client.ClientHttpRequestExecution;
-import org.springframework.http.client.ClientHttpRequestInterceptor;
-import org.springframework.http.client.ClientHttpResponse;
-
-import java.io.IOException;
-
-public class UserContextInterceptor implements ClientHttpRequestInterceptor {
-
-    @Override
-    public ClientHttpResponse intercept(HttpRequest httpRequest, byte[] bytes, ClientHttpRequestExecution clientHttpRequestExecution) throws IOException {
-
-        HttpHeaders httpHeaders = httpRequest.getHeaders();
-        httpHeaders.add(UserContext.CORRELATION_ID, UserContextHolder.getContext().getCorrelationId());
-        httpHeaders.add(UserContext.AUTH_TOKEN, UserContextHolder.getContext().getAuthToken());
-
-        return clientHttpRequestExecution.execute(httpRequest, bytes);
-    }
-}
